pyproject.toml
[project]
name = "podservice"
version = "0.1.0"
description = "YouTube to Podcast Feed Service"
readme = "README.md"
authors = [{name = "Ivan Kovnatsky"}]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "flask",
    "watchdog",
    "yt-dlp",
    "PyYAML",
    "requests",
    "click",
    "Pillow",
]

[project.urls]
Homepage = "https://github.com/ivankovnatsky/podservice"

[project.scripts]
podservice = "podservice.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "podservice"
version = "0.1.0"
description = "YouTube to Podcast Feed Service"
authors = ["Ivan Kovnatsky"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/ivankovnatsky/podservice"

[tool.poetry.dependencies]
python = ">=3.8"
flask = "*"
watchdog = "*"
yt-dlp = "*"
PyYAML = "*"
requests = "*"
click = "*"
Pillow = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[tool.poetry.scripts]
podservice = "podservice.cli:main"

[tool.setuptools]
packages = ["podservice"]

[tool.ruff.lint]
extend-select = ["I"]
